Add base class for dialog fragments
