+++ android/java/org/chromium/chrome/browser/vpn/BraveVpnCalloutDialogFragment.java
@@ -7,50 +7,19 @@
 
 package org.chromium.chrome.browser.vpn;
 
-import android.content.DialogInterface;
-import android.content.res.Configuration;
 import android.os.Bundle;
-import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
-
-import androidx.fragment.app.DialogFragment;
 
 import org.chromium.chrome.R;
-import org.chromium.chrome.browser.util.ConfigurationUtils;
 import org.chromium.chrome.browser.vpn.BraveVpnPrefUtils;
 import org.chromium.chrome.browser.vpn.BraveVpnUtils;
-import org.chromium.ui.base.DeviceFormFactor;
 
-public class BraveVpnCalloutDialogFragment extends DialogFragment implements View.OnClickListener {
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        setDialogParams();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        getDialog().setOnKeyListener(new DialogInterface.OnKeyListener() {
-            @Override
-            public boolean onKey(android.content.DialogInterface dialog, int keyCode,
-                    android.view.KeyEvent event) {
-                if ((keyCode == android.view.KeyEvent.KEYCODE_BACK)) {
-                    dismiss();
-                    return true;
-                } else
-                    return false;
-            }
-        });
-        setDialogParams();
-    }
-
+public class BraveVpnCalloutDialogFragment
+        extends BraveVpnDialogFragment implements View.OnClickListener {
     @Override
     public View onCreateView(
             LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -74,23 +43,5 @@ public class BraveVpnCalloutDialogFragment extends DialogFragment implements Vie
             BraveVpnUtils.openBraveVpnPlansActivity(getActivity());
         }
         dismiss();
-    }
-
-    private void setDialogParams() {
-        DisplayMetrics displayMetrics = new DisplayMetrics();
-        getActivity().getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
-        int mDeviceHeight = displayMetrics.heightPixels;
-        int mDeviceWidth = displayMetrics.widthPixels;
-
-        WindowManager.LayoutParams params = getDialog().getWindow().getAttributes();
-        boolean isTablet = DeviceFormFactor.isNonMultiDisplayContextOnTablet(getActivity());
-        boolean isLandscape = ConfigurationUtils.isLandscape(getActivity());
-        if (isTablet || isLandscape) {
-            params.width = (int) (0.5 * mDeviceWidth);
-        } else {
-            params.width = (int) (0.9 * mDeviceWidth);
-        }
-        params.height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        getDialog().getWindow().setAttributes(params);
     }
 }
