+++ android/java/org/chromium/chrome/browser/vpn/BraveVpnConfirmDialogFragment.java
@@ -8,49 +8,17 @@
 package org.chromium.chrome.browser.vpn;
 
 import android.animation.Animator;
-import android.content.DialogInterface;
-import android.content.res.Configuration;
 import android.os.Bundle;
-import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.WindowManager;
-import android.widget.LinearLayout;
-
-import androidx.fragment.app.DialogFragment;
 
 import com.airbnb.lottie.LottieAnimationView;
 
 import org.chromium.chrome.R;
-import org.chromium.chrome.browser.util.ConfigurationUtils;
-import org.chromium.ui.base.DeviceFormFactor;
 
-public class BraveVpnConfirmDialogFragment extends DialogFragment {
+public class BraveVpnConfirmDialogFragment extends BraveVpnDialogFragment {
     private LottieAnimationView mAnimatedView;
-
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        setDialogParams();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        getDialog().setOnKeyListener(new DialogInterface.OnKeyListener() {
-            @Override
-            public boolean onKey(android.content.DialogInterface dialog, int keyCode,
-                    android.view.KeyEvent event) {
-                if ((keyCode == android.view.KeyEvent.KEYCODE_BACK)) {
-                    dismiss();
-                    return true;
-                } else
-                    return false;
-            }
-        });
-        setDialogParams();
-    }
 
     @Override
     public View onCreateView(
@@ -83,27 +51,5 @@ public class BraveVpnConfirmDialogFragment extends DialogFragment {
             public void onAnimationRepeat(Animator animation) {}
         });
         mAnimatedView.playAnimation();
-    }
-
-    private void setDialogParams() {
-        DisplayMetrics displayMetrics = new DisplayMetrics();
-        getActivity().getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
-        int mDeviceHeight = displayMetrics.heightPixels;
-        int mDeviceWidth = displayMetrics.widthPixels;
-
-        WindowManager.LayoutParams params = getDialog().getWindow().getAttributes();
-        boolean isTablet = DeviceFormFactor.isNonMultiDisplayContextOnTablet(getActivity());
-        boolean isLandscape = ConfigurationUtils.isLandscape(getActivity());
-        if (isTablet) {
-            params.width = (int) (0.5 * mDeviceWidth);
-        } else {
-            if (isLandscape) {
-                params.width = (int) (0.5 * mDeviceWidth);
-            } else {
-                params.width = (int) (0.9 * mDeviceWidth);
-            }
-        }
-        params.height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        getDialog().getWindow().setAttributes(params);
     }
 }
